Mg/reader xrcs update (#327)
* now assigning error as coordinate intead of attribute

* Updated helike_spectroscopy to get latest MDS+ structure

* new XRCS quantities added

* new XRCS quantities added

* filter_coords no longer reordering dims / error added as coordinate of data rather than attribute

* precommit

* changed spectra to intens

* defaults split into save and load (removes issues with circular imports)

* updated with new MDS+ quantities

* precommit

* precommit

* read_write_default objects -> load_default_objects

* read_write_default objects -> load_default_objects

* xrcs updated quantities

* fixing import error

* precommit

* helike spectroscopy gets spectra and raw_spectra for uncalibrated signal

* helike spectroscopy gets spectra and raw_spectra for uncalibrated signal

* filter names updated

* pre-commit

* helike spectroscopy gets spectra and raw_spectra for uncalibrated signal
