+++ indica/datatypes.py
@@ -29,6 +29,9 @@ UNITS: Dict[str, str] = {
     "density_integrated": "1/$m^2$",
     "brightness": "$W/m^2$",
     "emissivity": "$W/m^3$",
+    "intensity": "$count/s$",
+    "radiance": "W/m^2/steradian$",
+    "spectral_radiance": "W/m^2/steradian/nm$",
     "velocity": "m/s",
     "magnetic_flux": r"$Wb/2\pi$",
     "f": "Wb m",
@@ -88,6 +91,9 @@ DATATYPES: Dict[str, Tuple[str, str]] = {
         "Centrifugal asymmetry multiplier",
         "none",
     ),
+    "intensity": ("Intensity", "intensity"),
+    "radiance": ("Radiance", "radiance"),
+    "spectral_radiance": ("Spectral Radiance", "spectral_radiance"),
     "emissivity": ("Emissivity", "emissivity"),
     "total_radiated_power_emission": ("$P_{rad, tot}$", "emissivity"),
     "sxr_radiated_power_emission": ("$P_{rad, sxr}$", "emissivity"),
