+++ indica/readers/available_quantities.py
@@ -27,15 +27,17 @@ AVAILABLE_QUANTITIES: Dict[str, Dict[str, str]] = {
         "zeff": "effective_charge",
     },
     "get_helike_spectroscopy": {
+        "ti_w": "ion_temperature",
+        "ti_z": "ion_temperature",
+        "te_n3w": "electron_temperature",
+        "te_kw": "electron_temperature",
         "int_w": "line_intensity",
         "int_k": "line_intensity",
         "int_tot": "line_intensity",
         "int_n3": "line_intensity",
-        "te_kw": "electron_temperature",
-        "te_n3w": "electron_temperature",
-        "ti_w": "ion_temperature",
-        "ti_z": "ion_temperature",
-        "spectra": "spectra",
+        "raw_spectra": "intensity",
+        "spectra": "spectral_radiance",
+        "background": "intensity",
     },
     "get_ppts": {
         "ne_rho": "electron_density",
