+++ indica/readers/read_st40.py
@@ -39,7 +39,14 @@ FILTER_LIMITS: Dict[str, Dict[str, tuple]] = {
     "cxff_pi": {"ti": (0, np.inf), "vtor": (0, np.inf)},
     "cxff_tws_c": {"ti": (0, np.inf), "vtor": (0, np.inf)},
     "cxqf_tws_c": {"ti": (0, np.inf), "vtor": (0, np.inf)},
-    "xrcs": {"ti_w": (0, np.inf), "te_kw": (0, np.inf), "te_n3w": (0, np.inf)},
+    "xrcs": {
+        "ti_w": (0, np.inf),
+        "ti_z": (0, np.inf),
+        "te_kw": (0, np.inf),
+        "te_n3w": (0, np.inf),
+        "spectra": (0, np.inf),
+        "raw_spectra": (0, np.inf),
+    },
     "brems": {"brightness": (0, np.inf)},
     "halpha": {"brightness": (0, np.inf)},
     "sxr_spd": {"brightness": (0, np.inf)},
@@ -66,6 +73,7 @@ FILTER_COORDS: Dict[str, Dict[str, tuple]] = {
     "cxff_tws_c": {"ti": ("channel", (0, np.inf)), "vtor": ("channel", (0, np.inf))},
     "xrcs": {
         "spectra": ("wavelength", (0.0, np.inf)),
+        "raw_spectra": ("wavelength", (0.0, np.inf)),
     },
     "ts": {"te": ("channel", (0, np.inf)), "ne": ("channel", (0, np.inf))},
 }
@@ -300,17 +308,17 @@ class ReadST40:
         self.get_equilibrium(R_shift=R_shift, revision=revisions["efit"])
         for instrument in instruments:
             print(f"Reading {instrument}")
-            try:
-                self.get_raw_data(
-                    "",
-                    instrument,
-                    revisions[instrument],
-                    set_equilibrium=set_equilibrium,
-                )
-            except Exception as e:
-                print(f"error reading: {instrument} \nException: {e}")
-                if debug:
-                    raise e
+            # try:
+            self.get_raw_data(
+                "",
+                instrument,
+                revisions[instrument],
+                set_equilibrium=set_equilibrium,
+            )
+            # except Exception as e:
+            #     print(f"error reading: {instrument} \nException: {e}")
+            #     if debug:
+            #         raise e
 
         if raw_only:
             return
@@ -360,7 +368,14 @@ def bin_data_in_time(
 
             if "t" in data_quant.coords:
                 data_quant = convert_in_time_dt(tstart, tend, dt, data_quant)
-
+            # Using groupedby_bins always removes error from coords so adding it back
+            if "error" in raw_data[instr][quant].coords:
+                error = convert_in_time_dt(
+                    tstart, tend, dt, raw_data[instr][quant].error
+                )
+                data_quant = data_quant.assign_coords(
+                    error=(raw_data[instr][quant].dims, error)
+                )
             binned_quantities[quant] = data_quant
         binned_data[instr] = binned_quantities
     return binned_data
@@ -408,7 +423,7 @@ def coord_condition(data: DataArray, coord_info: tuple):
     condition = (data.coords[coord_name] >= coord_slice[0]) * (
         data.coords[coord_name] < coord_slice[1]
     )
-    filtered_data = xr.where(condition, data, np.nan)
+    filtered_data = data.where(condition, np.nan)
     filtered_data.attrs = data.attrs
     return filtered_data
